refactor: repository가 일급 컬렉션을 반환하도록 변경
