+++ src/main/resources/application.yml
@@ -10,9 +10,6 @@ spring:
       path: /h2-console
   jpa:
     show-sql: true
-    properties:
-      hibernate:
-        format_sql: true
     hibernate:
       ddl-auto: create-drop
     defer-datasource-initialization: true
